+++ Documentation/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Set Loader editor only to not prevent VRChat upload
 - Cache created recording files
+- Store takes as an array, better name handling
 
 ## [2.0.0] - 2026-09-08